requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2
chromadb
pdf2image
